Extract program counter from Processor
The Processor does not need to know the program counter. This is
something decided externally and given to it when it has to
interpret an instruction.

However, the processor may now how to update the program counter
to find the next instruction so it returns a program counter
updater for that.